Move the "elements" into components
Also broke the map demo page
