+++ src/ui/components/slay-the-web-text.js
@@ -1,9 +1,9 @@
 import {html, render} from '../lib.js'
 import createNewGame from '../../game/new-game.js'
 import {getCurrRoom} from '../../game/utils-state.js'
-import Cards from '../components/cards.js'
-import {SlayMap} from '../components/map.js'
-import {Monster, Player} from '../components/player.js'
+import Cards from './cards.js'
+import {SlayMap} from './map.js'
+import {Monster, Player} from './player.js'
 // import CardChooser from './card-chooser.js'
 // import CampfireRoom from './campfire.js'
 // import StartRoom from './start-room.js'
@@ -145,4 +145,6 @@ export default class SlayTheWebText extends HTMLElement {
 	}
 }
 
-customElements.define('slay-the-web-text', SlayTheWebText)
+if (!customElements.get('slay-the-web-text')) {
+	customElements.define('slay-the-web-text', SlayTheWebText)
+}
