+++ src/ui/env.d.ts
@@ -1,3 +1,3 @@
 /// <reference path="../../.astro/types.d.ts" />
-/// <reference path="../.astro/types.d.ts" />
 /// <reference types="astro/client" />
+/// <reference path="../.astro/types.d.ts" />
