+++ src/ui/pages/index.astro
@@ -2,10 +2,8 @@
 import Layout from '../layouts/Layout.astro'
 import '../styles/index.css'
 ---
-<script src="../elements/slay-the-web.js"></script>
+<script src="../components/slay-the-web.js"></script>
 
 <Layout title="Slay the Web">
   <slay-the-web></slay-the-web>
 </Layout>
-
-<style></style>
