+++ src/ui/pages/text.astro
@@ -2,7 +2,7 @@
 import Layout from '../layouts/Layout.astro'
 // import '../styles/index.css'
 ---
-<script src="../elements/slay-the-web-text.js"></script>
+<script src="../components/slay-the-web-text.js"></script>
 
 <Layout title="Slay the Web: text version">
   <slay-the-web-text></slay-the-web-text>
